Add specs for Bulk Uploader
Per discussions with @slifty and @kfogel, we would like to start
documenting certain product specifications here in the meta repo and not
follow the anti-pattern of documenting them as “umbrella issues” that
never get closed.[1]

This commit adds the spec for the Bulk Uploader redesign. @slifty and I
chose this directory structure together.

[1] https://github.com/PhilanthropyDataCommons/front-end/issues/1027